+++ core-database/src/main/java/android/template/core/database/Trip.kt
@@ -11,10 +11,10 @@ import kotlinx.coroutines.flow.Flow
 @Entity
 data class Trip(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0L,
     val startTimestamp: Long,
     val duration: Long,
-    val distance: Double
+    val distance: Int
 )
 
 @Dao
@@ -31,5 +31,5 @@ interface TripDao {
     @Insert
     suspend fun insertTrip(
         item: Trip
-    ): Int
+    ): Long
 }
